Sprite updates and smaller code alterations

+++ Items/Weapons/RubySword.cs
@@ -29,7 +29,7 @@ public class RubySword : ModItem
     {
         if (Main.rand.NextBool(10))
         {
-            int dust = Dust.NewDust(new Vector2((float)hitbox.X, (float)hitbox.Y), hitbox.Width, hitbox.Height, DustID.Firework_Red, 0f, 0f, 100, default, 2f);
+            int dust = Dust.NewDust(new Vector2(hitbox.X, hitbox.Y), hitbox.Width, hitbox.Height, DustID.Firework_Red, 0f, 0f, 100, default, 2f);
             Main.dust[dust].noGravity = true;
         }
     }
